// lib/api/generated/models/HeaderVideoAsset.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type HeaderVideoAsset = {
    file_id?: (string | null);
    name?: (string | null);
    url: string;
    thumbnail_url?: (string | null);
    file_path?: (string | null);
    mime?: (string | null);
    size?: (number | null);
};